+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Enhancements:
+
+- Fix `AimLogger` deprecation issues related to release of PyTorch Lightning v1.7 (djwessel)
+
 ## 3.13.0 Aug 21, 2022
 
 ### Enhancements: